Add dependency: flake8-docstrings
flake8-docstrings is "a simple module that adds an extension for the fantastic
pep257 tool to flake8." It helps ensure that docstrings comply with PEP 257.
See:

* https://gitlab.com/pycqa/flake8-docstrings
* http://pep257.readthedocs.org/en/latest/

The flake8-pep257 plug-in was also considered, but Robpol86/flake8-pep257#4 was
encountered. In addition, flake8-docstrings appears to be compatible with a
greater variety of Python versions.

Targets #22.