WIP: Skip known failing tests
Skip failing tests. This is a temporary solution until #29 and #30 are fixed.
The idea is that it's better to highlight real regressions than drown the user
in known issues. In the immediate future, developers who'd like to know if
they've fixed an issue can remove relevant `skipTest()` lines from their local
copies of the test suite. Not optimal, but workable.

Also, update a test that made an assertion about an undesirable behaviour.
Method `ReadUpdateDeleteTestCase.test_deleted_user` in module
`pulp_smash.tests.platform.api_v2.test_user` now correctly passes a JSON body in
a PUT request for a user.

Tested against Pulp 2.6 and 2.7:

    $ python -m unittest2 discover pulp_smash.tests
    ....s...sssssss.s.....ssss........................................s........
    ----------------------------------------------------------------------
    Ran 72 tests in 30.678s

    OK (skipped=14)