Force the use of older linters
Pylint 1.5.0 and astroid 1.4.1 were released over the past weekend.
Unfortunately, there are several regressions in the new releases. One such
regression is a `UnicodeEncodeError` on Python 2, as documented here:
https://bitbucket.org/logilab/astroid/issues/273

The regression has since been fixed, but no release has been made that includes
the fix. Until the fix is made available via a release, force the use of older
versions of pylint and astroid.

Before this change:

    $ pip freeze | grep 'astroid\|pylint'
    astroid==1.4.1
    pylint==1.5.0

After this change:

    $ pip freeze | grep 'astroid\|pylint'
    astroid==1.3.8
    pylint==1.4.4

Related to #46.