Add decorator for conditionally skipping tests
Fix #29, "Add mechanism to select tests based on Pulp version":

> Some tests are only applicable to a particular Python version. Let's add some
> mechanism that allows users to run only tests which apply to a particular Pulp
> version or a range of Pulp versions.

A new "version" attribute was recently added to
`pulp_smash.config.ServerConfig`. Build on top of that work by adding a new
decorator, `pulp_smash.utils.requires`. As its docstring states:

    This decorator concisely encapsulates a common pattern for skipping tests.
    It can be used like so:

    >>> from pulp_smash.config import get_config
    >>> from pulp_smash.utils import require
    >>> from unittest import TestCase
    >>> class MyTestCase(TestCase):
    ...
    ...     @classmethod
    ...     def setUpClass(cls):
    ...         cls.cfg = get_config()
    ...
    ...     @require('2.7')  # References `self.cfg`
    ...     def test_foo(self):
    ...         pass  # Add a test for Pulp 2.7+ here.

    Notice that ``cls.cfg`` is assigned to. This is a **requirement**.

Of special note is that the decorator will throw an exception if an invalid
version is passed. For example, this will throw an exception:

    @require('2.7.foobar')
    def test_foo(self):
        pass

Many thanks to @bmbouter for help with this change.