Add a mechanism to skip known failing tests
Fix #30:

> Let's add some mechanism that allows us to skip tests that are known to fail
> on certain versions of Pulp.

Why would this be useful? Well, it's very handy to be able to add tests for a
bug as soon as a bug is filed. However, these tests will necessarily fail for a
while before the bug is fixed and available in Pulp. And we can't just skip that
failing test with the `@require` decorator, as it's unknown which version of
Pulp the bug will be fixed in.

Add a pair of methods:

* `pulp_smash.utils.bug_is_testable`
* `pulp_smash.utils.bug_is_untestable`

These methods can be used in a test method like so:

    if bug_is_untestable(1356):
        self.skipTest('https://pulp.plan.io/issues/1356')

Add unit tests for these two new methods. These methods are resilient against:

* …malformed inputs. If anything other than an integer is passed in, an
  exception is thrown. This allows the methods to cache bug statuses.
* …connection errors. If the Pulp bug tracker cannot be contacted, Pulp Smash
  will issue a warning and assume that the bug is testable.