Extend timeout when waiting for tasks to complete
From #62:

> on some test runs, every polled task (or at least most polled tasks) will fail
> to complete. As a result, you'll see results like "37 failed tests". It's
> unclear whether we get catastrophic test failures because of errors somewhere
> (in Pulp, or the system hosting it, or...), or simply because the systems
> under test are slow.

An ideal solution is to add some mechanism to Pulp Smash or its test runner so
that the default timeout for tasks can be configured. However, a much more
pragmatic solution at this time is simply to extend the hard-coded time-out from
50 to 120 seconds. Do so.

Test suite results do not change as a result of this commit. Results:

    ============  =============================
    Pulp Version  Test Suite Results
    ============  =============================
    2.7           OK (skipped=14)
    dev (2.8)     FAILED (errors=2, skipped=16)
    ============  =============================