Change variable used by tearDownClass
Update class `_BaseTestCase` in module
`pulp_smash.tests.rpm.api_v2.test_iso_crud`. Drop the tuple named "hrefs" in
favor of a set named "resources." Tuples are immutable, which makes them a bad
fit for a variable which may be continually updated throughout a `setUpClass`
method; and the new name better reflects what the variable is used for.

Test suite results do not change as a result of this commit:

    ============  =============================
    Pulp Version  Test Suite Results
    ============  =============================
    2.7           OK (skipped=14)
    dev (2.8)     FAILED (errors=2, skipped=16)
    ============  =============================