Add a CLI helper
Add, document and test module `pulp_smash.cli`. This module provides "tools for
working with Pulp's CLI." This fixes #32:

> It'd be nice if we could inspect the state of remote systems under test. For
> example, it may be useful to log in to a remote system and determine whether
> certain files are lying around in certain places in the file system.

Here's a simple example of how to use the new module:

    >>> from pulp_smash import cli, config
    >>> server_config = config.ServerConfig('localhost')
    >>> client = cli.Client(server_config)
    >>> response = client.run(('echo', '-n', 'foo'))
    >>> response.returncode == 0
    True
    >>> response.stdout == 'foo'
    True
    >>> response.stderr == ''
    True

The `Client` class is flexible enough that both request and response handling
can be customized on a per-instance basis. See the docstrings in module
`pulp_smash.cli` for details.

Add "plumbum" as a dependency. It is used by the CLI helper.

Many thanks to @elyezer for help in brainstorming a solution to this issue.

This commit does not affect test results (107 tests):

    ============  =============================
    Pulp Version  Test Suite Results
    ============  =============================
    2.7           OK
    dev (2.8)     FAILED (errors=2, skipped=13)
    ============  =============================