Refactor RPM repository API tests (2/2)
Rewrite module `pulp_smash.tests.rpm.api_v2.test_iso_crud` to use the new
`pulp_smash.api.Client` API client. Doing this makes the module more readable
and concise.

As a result of this change, one formerly unknown bug has been discovered in the
development version of Pulp. This bug was already present, but was not exposed
to due a bug in the test code. The test code had this code:

    if response.status_code == 202:
        utils.poll_spawned_tasks(server_config, response.json())

`poll_spawned_tasks` is a generator, so the above code did not poll all tasks,
and therefor did not discover that one of the tasks consistently throws an HTTP
500 error. The new test code automatically evaluates the generator.

The updated test module produces the following results (11 and 7 tests,
repectively):

    ============  ==================
    Pulp Version  Test Suite Results
    ============  ==================
    2.7           OK (skipped=1)
    dev (2.8)     FAILED (errors=1)
    ============  ==================

Related to #64.