Refactor and test helper function
Refactor function `pulp_smash.config._get_config_file_path`. Make the function
calculate a set of paths just once instead of twice.

Update the wording of the exception raised by this function. Notably, this
function now states which paths are searched if it fails to find a file, so
there is no need to redundantly state the name of the file it is searching for.
Here's an example of how to trigger the exception and what the message looks
like:

    >>> from pulp_smash import config
    >>> cfg = config.ServerConfig().read()
    …
    pulp_smash.exceptions.ConfigFileNotFoundError: Pulp Smash is unable to find a configuration file. The following (XDG compliant) paths have been searched: /home/ichimonji10/.config/pulp_smash/settings.json

Add a test case for this function.