Do not measure coverage for unreachable code paths
Consider the following block of code:

    try:  # try Python 3 import first
        from urllib.parse import urljoin
    except ImportError:
        from urlparse import urljoin

If Python 2 executes this block of code, all four statements will be evaluated.
If Python 2 executes this block of code, only the first two statements will be
evaluated. As a result, Python 2 and Python 3 will give different unit test
coverage numbers. Address this issue by ignoring coverage for code branches that
are unreachable under a given version of Python. Results:

    ===========  ======  =====
    Python Ver.  Before  After
    ===========  ======  =====
    2            79%     78%
    3            77%     78%
    ===========  ======  =====