Add function for resetting Pulp
Add function `pulp_smash.utils.reset_pulp`. When executed, this function will:

> Stop Pulp, reset its database, remove certain files, and start it.

One of the existing tests in Pulp Smash logs in to a target system and resets
the Pulp services. As a result, both that test and this new function contain a
semantically identical set of service names. Deal with this issue by creating a
new constant, `pulp_smash.constants.PULP_SERVICES`.

Fix #5. Closely related to #80 — thanks to @dkliban for providing source
material for this commit in that pull request.