Make docker CLI tests log in during test setUp
pulp-admin talks to Pulp via its API. As a result, pulp-admin must supply
credentials with each request, which means that users must somehow supply
credentials to pulp-admin.

One way to handle this issue is to provide a username and password with each
request. Another option is to execute `pulp-admin login -u ${username} -p
${password}`, which causes pulp-admin to fetch a certificate and save it as
`~/.pulp/user-cert.pem`. This certificate is then used in subsequent requests if
no username and password are provided.

Make each of the docker CLI test cases log in to the server using `pulp-admin
login`. This should fix several test failures.

Test results do not change as a result of this commit.