Change how OSTree tests are selected
A function named `load_tests` is in module `pulp_smash.tests.ostree`. This
function is automatically called by the unittest (or unittest2) test runner
according to the load_tests protocol. [1] This works well, but few if any third
party test runners support the protocol. (To the best of my knowledge, only
nose2 supports the load_tests protocol, and this support is nominal.)

Many third party test runners add JUnit-style XML reporting. This reporting is
valuable, especially when combined with a tool like Jenkins. Unfortunately, it's
not so easy to get reports with the unittest test runner.

The conflict here is tha we want JUnit-style XML reporting *and* easy test
selection. There's a couple routes forward:

* Drop support for unittest (and unittest2) and adopt the test selection
  features of a third party test runner.
* Plow forward and put more work into figuring out how to make unittest produce
  reports.
* Don't use the load_tests protocol. Instead, add `setUpModule` functions that
  perform the same task into each individual module of tests.

This commit adopts the third route. This route ensures that Pulp Smash continues
to be compatible with the standard library's `unittest` module and allows
reports to be generated by third party tools. The main drawback is some level of
redundant code. (Options for reducing this redundancy may be available.)

Test results are as expected for the following two commands:

    python -m unittest2 discover pulp_smash.tests.ostree
    py.test pulp_smash/tests/ostree/