Move several functions to module pulp_smash.api
Module `pulp_smash.api` contains several tools for working with Pulp's API.
Curiously, it imports two critically important function from module
`pulp_smash.utils`:

* `poll_spawned_tasks`
* `poll_task`

This is curious because module `pulp_smash.utils` contains a variety of
utilities, such as the `reset_pulp` function which resets the database and cache
on a Pulp server. Given this, shouldn't `pulp_smash.utils` build on the client
provided by `pulp_smash.api`, and not the other way around?

Fix this by moving these functions to module `pulp_smash.api`, and update the
docstring in module `pulp_smash.utils` to clearly state the relationship between
the two modules.

This commit doesn't affect test results. Both before and after this commit:

    =========  ==========  ==================
    Pulp Ver.  Num. Tests  Test Suite Results
    =========  ==========  ==================
    2.7        138         OK (skipped=19)
    dev (2.8)  127         OK (skipped=13)
    =========  ==========  ==================

Test results generated with:

    python -m unittest2 discover pulp_smash.tests

Fix #110. That issue asks for various utilities that make use of the API and ClI
to be moved out of `pulp_smash.utils`, which this commit does not do. However,
that request was made primarily becuase of a circular import issue, and this
solves that underlying issue.