Add sync test of unnamespaced repo from Dockerhub
The `upstream-name` of a Docker repository indicates the correct feed
to sync from the registry. For Dockerhub only, `registry-1.docker.io`,
some special repositories do not have a namespace, any it would be
correct to set their upstream name to be the repo_name by itself. In
this case, the sync will work exactly the same as if the upstream
name were `library/<repo_name>`.

Newly added tests pass:

    (pulp-smash)[vagrant@dev pulp-smash]$ python -m unittest2 \
    > pulp_smash.tests.docker.cli.test_sync
    s...........
    ----------------------------------------------------------------------
    Ran 12 tests in 56.368s

Resolves #105

Prevents regression of the folloing Pulp issue:
https://pulp.plan.io/issues/1404