Correctly poll multi-task call reports
Function `pulp_smash.api.poll_task` nominally polls Pulp until a task is
complete, then recursively polls each of its children. Unfortunately, it
fails to correctly do this. It truly will wait until the given task
completes, but if that task has itself spawned child tasks, it will fail
to poll them and instead return a generator. Fix this.

This issue likely has not been an issue up to this point due to most or
all tasks spawning no children. This issue was discovered while
reviewing #117, which has tasks of this kind.

Test results do not change as a result of this commit.