Edit code samples in README.rst
Edit both of the code samples in the root readme file. The first code sample
incorrectly stated that Pulp Smash can be installed with `pip install
pulp_smash`, but the actual package name is `pulp-smash`. The second code sample
neglected to explain that source code must be present in order to locally
generate documentation.