Add utilities for docker CLI tests
Add and document module `pulp_smash.tests.docker.cli.utils`. This module
contains a variety of functions that may be useful when writing docker
CLI tests.

This module is unused, but it should be of use in the existing modules
in package `pulp_smash.tests.docker.cli`. In addition, it will be of use
in the soon-to-be-added tests described by #98.

These functions are derived from those in #126. Thank you, @rbarlow.