Fix bug in reset_pulp utility function
When the CLI client executes commands on a remote host, it does not
perform typical shell conveniences such as globbing or pipe redirection.
As a result, a command like:

    touch "$HOME"/*

…does not actually touch all files in the home directory, but instead
touches a file named "*".

Function `pulp_smash.utils.reset_pulp` attempts to use globbing. Fix it.
This change does not affect test suite results.