Use a single make target for all CI checks
Travis CI is our canonical source of information regarding whether a
build is sane. A user can perform approximately the same task by
executing various make targets. [1] However, this means executing six
different make targets. Though this is doable, it is not human friendly,
and this friction likely makes users less inclined to perform a full set
of checks locally before pushing changes to GitHub.

Fix this by creating a new make target, `all`, which runs checks in an
appropriate order. Make Travis use this target, and make the
documentation reference this target.

Much thought has been put into the ordering of the targets. Comments are
inline, and one of the comments mentions timing information. Following
are the timing measurements used:

    =============  =========================
    make target    execution times (seconds)
    =============  =========================
    docs-html      14.6, 16.4, 16.4
    package        1.8, 1.8, 1.8
    test           1.0, 1.0, 1.0
    test-coverage  1.3, 1.3, 1.4
    lint           31.3, 31.9, 33.4
    =============  =========================

Each make target was executed with a command in the following form:

    for i in {1..3}; do
        git clean -dfx && time make "$target"
    done

[1] Travis provides a pristine environment with several different Python
    interpreters. That's hard to reproduce as an end user.