Drop unnecessary test method
Drop an a test method that verifies that a CLI command status code is 0.
This assertion is unnecessary because function `repo_sync` in module
`pulp_smash.tests.docker.cli.utils` already does this.

Test results are sane after this change:

    =========  ==========  ==============
    Pulp Ver.  Num. Tests  Test Results
    =========  ==========  ==============
    2.7        ?           ?
    dev (2.8)  23          OK (skipped=1)
    =========  ==========  ==============

Test results generated with:

    python -m unittest2 pulp_smash.tests.docker.cli.test_sync_publish

Pulp 2.7 test results are not available at this time due to a packaging
error.  However, the most heavily affected test case,
`SyncPublishV2TestCase`, is correctly skipped with this command:

    python -m unittest2 \
    pulp_smash.tests.docker.cli.test_sync_publish.SyncPublishV2TestCase