Remove orphans in BaseAPITestCase.tearDownClass
Method `BaseAPITestCase.tearDownClass` iterates through a set of hrefs
and deletes each resource named therein. This approach works well when
the test author explicitly creates a small set of resources. However,
one common scenario is to create a repository and sync it. In this case,
additional resources are automatically created by Pulp, and it is
impossible to know what will be created until runtime. Even worse, when
the parent repository is deleted, orphan resources will be left behind.

Fix #158. Delete orphan resources at the end of
`BaseAPITestCase.tearDownClass`. One can verify that this change works
by issuing an HTTP GET to `pulp_smash.constants.ORPHANS_PATH` and
inspecting the results. Here's a snippet of results from a system after
this commit is applied:

    {
        u'distribution': {
            u'_href': u'/pulp/api/v2/content/orphans/distribution/',
            u'count': 0
        },
        u'docker_blob': {
            u'_href': u'/pulp/api/v2/content/orphans/docker_blob/',
            u'count': 0
        },
        u'docker_image': {
            u'_href': u'/pulp/api/v2/content/orphans/docker_image/',
            u'count': 0
        },

This commit does not affect test results.