Test relative_path validation in OSTree
Ensure that the relative path in the config of OSTree distributors is
validated against conflicts from the distributors of other repositories.
This should occur on creation and update of an OSTree distributor, and
should also protect against possible conflicts related to sub-urls.

Tests should confirm that https://pulp.plan.io/issues/1106 is fixed.