Respect target_platform_release bug tracker field
Change the signatures of `bug_is_testable` and `bug_is_untestable` in
module `pulp_smash.selectors`, from this:

    bug_is_testable(bug_id)
    bug_is_untestable(bug_id)

To this:

    bug_is_testable(bug_id, pulp_version)
    bug_is_untestable(bug_id, pulp_version)

The new `pulp_version` parameter is "A ``packaging.version.Version``
object telling the version of the Pulp server we are testing." It lets
the function determine whether a bug is testable or not based not only
on the state of a bug, but the Pulp version in which the bug was fixed
relative to the Pulp version we're testing against. Sample usage:

    >>> from packaging.version import Version
    >>> from pulp_smash.selectors import bug_is_testable
    >>> bug_is_testable(1743, Version('2.7.0'))
    False
    >>> bug_is_testable(1743, Version('2.7.99999'))
    False
    >>> bug_is_testable(1743, Version('2.8'))
    True
    >>> bug_is_testable(1743, Version('2.8.0'))
    True
    >>> bug_is_testable(1743, Version('2.8.0.0'))
    True
    >>> bug_is_testable(1743, Version('2.8.1'))
    True
    >>> bug_is_testable(1743, Version('3'))
    True

If a bug's "Target Platform Release" field is unset, the field is
effectively ignored. [1] For example:

    >>> from packaging.version import Version
    >>> from pulp_smash.selectors import bug_is_testable
    >>> bug_is_testable(1744, Version('2'))
    True
    >>> bug_is_testable(1744, Version('3'))
    True

Test suite results do not change as a result of this commit. Results
generated with:

    python -m unittest2 discover pulp_smash.tests

Fix #173.

[1] More precisely, we assume that the bug has been fixed in Pulp
    version 0.