Rewrite "Installation" documentation section
Do not attempt to lay out a comprehensive set of installation
strategies with wordy explanations. Instead, list a recipe for creating
a virtualenv, followed by three strategies for installing Pulp Smash,
followed by a link to the official Python documentation.

Drop a redundant set of instructions in the "About" page, and reference
the recipes in the "Installation" page instead.

This change is made on the theory that those people who need
instructions for installing a Python package will benefit from a concise
recipe-oriented approach, and those people who are interested in more
in-depth explanations of Python packaging are (should be!) looking
elsewhere.