Edit RPM API utility function get_repomd_xml
Change the method signature from this:

    get_repomd_xml(client, repo_url, repomd_type)

To this:

    get_repomd_xml(server_config, repo_path, repomd_type)

Passing in a server config instead of a client is likely less efficient,
as a `Client` object must be instantiated upon each function call.
However, this change has several positive effects. First, the method
implementation becomes dramatically simpler. Second, less object
mutation occurs: the method no longer needs to manage mutations to
``client.response_handler`` with a try/except block. Third, it is a
common pattern to pass a server config and path into a method, and
making this method do the same makes this method easier to learn.

Also, edit the function's docstring: fix a broken class reference, and
tweak the prose.

Test results do not change as a result of this change. Test results
generated with:

    python -m unittest2 \
        pulp_smash.tests.rpm.api_v2.test_comps_groups \
        pulp_smash.tests.rpm.api_v2.test_updateinfo