Add new dependency: python-dateutil
The ISO 8601 specification for dates and times is complex. It was first
introduced in 1988, and it has been updated in 2000 and 2004. It
specifies a broad variety of syntaxes. For example, the following can
all represent the same moment in time, at varying levels of exactness:

    2016-04-06
    2016-04-06T16:02:36+00:00
    2016-04-06T16:02:36Z
    20160406T160236Z
    2016-W14
    2016-W14-3
    2016-097

In addition, ISO 8601 allows for the specification of other values, such
as time intervals, repeating time intervals and time durations. As a
consequence, parsing ISO 8601 dates and times is not a trivial task, and
using a custom parser or hard-coded regexes seems unwise.

ISO 8601 is relevant because Pulp uses the specification for date and
time units. [1] In addition, module
`pulp_smash.tests.rpm.api_v2.test_export` parses ISO 8601 values, and
pull request 170 [2] also deals with such values.

python-dateutil is an appropriate choice as a new dependency because it
is widely used with ~232,000 downloads per day, it is compatible with a
wide variety of Python versions including 3.5, and it has been under
development since at least 2003.

Add python-dateutil as a required dependency. Use it in module
`pulp_smash.tests.rpm.api_v2.test_export`. Test results stay the same
after this change. Test results generated with:

    python -m unittest2 pulp_smash.tests.rpm.api_v2.test_export

[1] http://pulp.readthedocs.org/en/latest/user-guide/general-reference.html#date-and-time-units
[2] https://github.com/PulpQE/pulp-smash/pull/170