Refactor logic for per-plugin test skipping
Function `pulp_smash.utils.get_plugin_type_ids` is poorly named. It
doesn't return plugin IDs, as one might think, as this is impossible.
Instead, it returns the IDs of each content unit type currently
supported by a target Pulp server. In addition, this function doesn't
accept a `ServerConfig` object identifying a particular Pulp server to
target, but instead must work with the default `ServerConfig` object
returned by the global cache.

In addition, there is no utility function available which will raise a
`SkipTest` exception if the given content unit type is not supported.
Instead, individual modules must call `get_plugin_type_ids`, inspect the
response and raise their own exception. While not much of a hassle
individually, the fact that such a common task is not supported is
worrisome.

Fix these issues:

* Rename function `get_plugin_type_ids` to `get_unit_type_ids`. Make it
  require a `server_config` parameter.
* Create function `pulp_smash.utils.skip_if_type_is_unsupported`. Drop
  `pulp_smash.tests.ostree.utils.skip_if_no_plugin` in favor of this new
  function.
* Rewrite docstrings and update function usages as appropriate.
* Add unit tests for the new functions.

This change does not affect test results. Test suite run against a pair
of systems with the OSTree plugin installed and not, respectively, with
the following command:

    python -m unittest2 \
        pulp_smash.tests.ostree.api_v2.test_{crud,sync_publish}