Let linter check imports
The `lint-pylint` make target runs the Pylint linter over the code base.
A small number of checks are disabled, including import checks. Why are
import checks disabled? The import checks will issue a warning in the
case where four or more import lines are identical, and it will
therefore catch the following:

    try:  # try Python 3 first
        from urllib.parse import urljoin
    except ImportError:
        from urlparse import urljoin

This was considered a low-priority issue for some time. However, the
introduction of module `pulp_smash.compat` has solved this issue, and
there is now much less reason to ignore this check.

Update the `lint-pylint` make target and let Pylint check for
import-related issues.