Test whether updateinfo.xml suggests a reboot
If no value is specified for an errata's `reboot_suggested` field, the
errata is uploaded into a repository, and the repository is published,
the repository's `updateinfo.xml` file should not suggest a reboot. The
file's `<reboot_suggested>` tag should have a value of "False".

Add a test for: https://pulp.plan.io/issues/1782

Fix: https://github.com/PulpQE/pulp-smash/issues/214