Improve file upload logic in upload_import_unit()
`upload_import_unit` currently uploads a given file to Pulp all in one
go. This works well if the file is small, but is an issue when file
sizes become larger. Rewrite the function so that it iterates through
the given file and uploads the given file in 200,000 byte (~200 kB)
chunks.

This chunk size was arbitrarily chosen. It has been tested and works for
large files (specifically, a Docker image file) for Pulp servers running
on Fedora 22 and 23.

This change does not break existing tests. Tests executed with:

    python -m unittest2 discover pulp_smash.tests