Run children of SyncRepoBaseTestCase
Class `SyncRepoBaseTestCase` is an abstract base class, and it should
not run. However, children of this class are expected to implement
abstract functionality and should run. The current attempt to implement
this is as follows:

    @unittest2.skip(…)
    class SyncRepoBaseTestCase(…):

This fails, of course: child classes also execute this decorator. Use
reflection instead: a `SkipTest` exception is raised only if the first
element in the MRO tree is `SyncRepoBaseTestCase`.

This change allows more tests to run, without introducing any test
failures. Tests executed with:

    python -m unittest2 pulp_smash.tests.rpm.api_v2.test_sync_publish