Fix broken sudo logic
Function `utils.is_root` tells whether the Pulp Smash CLI client has
root access on a target system. A common use case for this function is
deciding whether to prefix a command with `sudo` or not. (`sudo` is an
optional system utility.) Fix an incorrect bit of logic that prepends
CLI commands with "sudo" when the test already has root access.