Update content unit counts tests
When one reads an RPM repository, the `content_unit_counts` attribute
lists how many units of each type are present in the repository.  Pulp
2.9 introduces support for the `package_langpacks` content type.  Update
a test that verifies the values in a repository's `content_unit_counts`
attribute.

The tests originally walked through each of the content types one at a
time and verified only specific content types. This works, but it means
that new content types can be added and the test will not notice. Make
the test more strict.

Related to: https://github.com/PulpQE/pulp-smash/issues/233