Refactor DuplicateUploadsMixin base test case
The `DuplicateUploadsMixin` is designed to test the following procedure:

1. Create a repository without a feed.
2. Upload a package into the repository.
3. Upload the same package into the repository.

The class is currently designed such that the two uploads should occur
in method `setUpClass`.

Refactor the test case so that repository creation occurs in
`setUpClass` and the uploads occur in the test methods. This makes it
easier to tell which of the two uploads failed, and it ensures the class
always cleans up after itself.

This change doesn't affect test coverage or introduce any test failures.
It adds four new tests. Tests executed with:

    python -m unittest2 \
        pulp_smash.tests.docker.api_v2.test_duplicate_uploads \
        pulp_smash.tests.puppet.api_v2.test_duplicate_uploads \
        pulp_smash.tests.python.api_v2.test_duplicate_uploads \
        pulp_smash.tests.rpm.api_v2.test_duplicate_uploads