Rewrite just-added tests for packages_directory
Rewrite the tests added by the previous commit. The new tests verify the
RPM plugin's support for a new distributor option, `packages_directory`.
(This new option is outlined in Pulp issue 1976. [1]) In addition, add
the new tests to the documentation.

The tests are in `PackagesDirectoryTestCase`, in module
`pulp_smash.tests.rpm.api_v2.test_packages_directory`. The test case
creates an RPM repository with a feed and syncs in packages, and then
publishes that repository several times with varying distributors and
varying publish options.

The test for Pulp's default behaviour (when `packages_directory` is not
specified) always executes. The other test cases skip based on the state
of Pulp issue 1976.

[1] https://pulp.plan.io/issues/1976