Update documentation in readme and docs/index.rst
Pulp smash has a presence on ReadTheDocs, PyPi and GitHub. Ensure that
both the readme and `docs/index.rst` link to all three of these sites.
This ensures that readers can find all of the resources they need, no
matter which website they find themselves on, or even if they just have
the source code.

Drop the instructions for building from source. A user who wants to do
something like that is also probably adventurous enough to run the
various makefiles to see what they can do. In addition, it seems strange
to explain how to generate documentation from source, but not explain
how to generate a Python package.